adjust floating ip logic

+++ scripts/kvm/create_haproxy_conf.sh
@@ -85,5 +85,5 @@ EOF
 done
 
 haproxy_pid=$(cat $lb_dir/haproxy.pid)
-kill -HUP $haproxy_pid
+[ $haproxy_pid -gt 0 ] && kill -HUP $haproxy_pid
 [ $? -ne 0 ] && ip netns exec $router haproxy -D -p $lb_dir/haproxy.pid -f $lb_dir/haproxy.conf
